requirements.txt
Flask==2.3.3
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
python-barcode==0.15.1
pyzbar==0.1.9